requirements.txt
streamlit>=1.28.0
pandas
plotly
numpy
sympy
qrcode[pil]
pyzbar
opencv-python-headless
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.0.0
scipy>=1.7.0
sympy>=1.8.0
pycryptodome
